backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
torch
torchvision
Pillow
numpy
python-multipart